@@ -41,11 +41,11 @@
     "@typescript-eslint/parser": "^5.49.0",
     "del-cli": "^4.0.1",
     "esbuild": "^0.14.54",
-    "eslint": "^8.32.0",
+    "eslint": "^8.33.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
     "npm-run-all": "^4.1.5",
-    "pnpm": "^7.26.1",
+    "pnpm": "^7.26.2",
     "prettier": "^2.8.3",
     "sass": "^1.57.1",
     "typescript": "^4.9.4",
